Added version to all test fixture classes and fixed GraphApiClientTest.PostVertexAsync_ShouldThrow_WhenVertexCollectionIsNotFound()
