+++ arangodb-net-standard.Test/AdminApi/AdminApiClientTestFixture.cs
@@ -21,6 +21,7 @@ namespace ArangoDBNetStandardTest.AdminApi
             await CreateDatabase(dbName);
             Console.WriteLine("Database " + dbName + " created successfully");
             ArangoDBClient = GetArangoDBClient(dbName);
+            await GetVersionAsync(ArangoDBClient);
             try
             {
                 var dbRes = await ArangoDBClient.Database.GetCurrentDatabaseInfoAsync();
