+++ arangodb-net-standard.Test/AnalyzerApi/AnalyzerApiClientTestFixture.cs
@@ -21,6 +21,7 @@ namespace ArangoDBNetStandardTest.AnalyzerApi
             await CreateDatabase(dbName);
             Console.WriteLine("Database " + dbName + " created successfully");
             ArangoDBClient = GetArangoDBClient(dbName);
+            await GetVersionAsync(ArangoDBClient);
             try
             {
                 var dbRes = await ArangoDBClient.Database.GetCurrentDatabaseInfoAsync();
